pyproject.toml
[project]
name = "voice-agnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "websocket>=0.2.1",
    "websockets>=15.0.1",
]
